[ENH]: Use new log service (#1971)
## Description of changes
*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - Use refactored log service

## Test plan
*How are these changes tested?*
Tested with tilt

+++ go/database/log/db/models.go
@@ -15,6 +15,6 @@ type Collection struct {
 type RecordLog struct {
 	Offset       int64
 	CollectionID string
-	Timestamp    int32
+	Timestamp    int64
 	Record       []byte
 }
